savegame/bson: merge writing utils to trx
