+++ src/libtrx/game/savegame/bson.c
@@ -1,17 +1,133 @@
 #include "game/savegame/bson.h"
 
 #include "bson.h"
+#include "game/game.h"
+#include "game/game_flow.h"
 #include "game/savegame.h"
+#include "game/shell.h"
 #include "log.h"
 #include "memory.h"
 #include "utils.h"
+#include "version.h"
 
 #include <zconf.h>
 #include <zlib.h>
 
 #define M_MAGIC_TR1X MKTAG('T', '1', 'M', 'B')
 #define M_MAGIC_TR2X MKTAG('T', '2', 'X', 'B')
-#define M_MAGIC_TRX MKTAG('T', 'R', 'X', 'S')
+#define M_MAGIC_TRX MKTAG('T', 'R', 'X', 'S') // TODO: use this after TRX 1.0
+
+static const char *M_GetSaveFilePattern(void);
+static void M_SaveToFile(MYFILE *fp, SAVEGAME_INFO *info);
+static bool M_LoadFromFile(MYFILE *fp);
+static bool M_LoadOnlyResumeInfo(MYFILE *fp);
+static bool M_UpdateDeathCounters(
+    MYFILE *fp, int32_t level_num, int32_t death_count);
+
+static SAVEGAME_STRATEGY m_Strategy = {
+    .allow_load = true,
+    .allow_save = true,
+    .format = SAVEGAME_FORMAT_BSON,
+    .get_save_file_pattern_func = M_GetSaveFilePattern,
+    .fill_info_func = Savegame_BSON_FillInfo,
+    .load_from_file_func = M_LoadFromFile,
+    .save_to_file_func = M_SaveToFile,
+    .load_only_resume_info_func = M_LoadOnlyResumeInfo,
+    .update_death_counters_func = M_UpdateDeathCounters,
+};
+
+static void M_SaveRaw(
+    MYFILE *const fp, const JSON_VALUE *const root, const int32_t level_num)
+{
+    size_t uncompressed_size;
+    char *uncompressed = BSON_Write(root, &uncompressed_size);
+
+    uLongf compressed_size = compressBound(uncompressed_size);
+    char *compressed = Memory_Alloc(compressed_size);
+    const int32_t result = compress(
+        (Bytef *)compressed, &compressed_size, (const Bytef *)uncompressed,
+        (uLongf)uncompressed_size);
+    if (result != Z_OK) {
+        Shell_ExitSystem("Failed to compress savegame data");
+    }
+
+    const GF_LEVEL *const level = GF_GetLevel(GFLT_MAIN, level_num);
+    const SAVEGAME_BSON_HEADER header = {
+        .magic = g_TRVersion == 1 ? M_MAGIC_TR1X : M_MAGIC_TR2X,
+        .initial_version = Savegame_GetInitialVersion(),
+        .version = SAVEGAME_CURRENT_VERSION,
+        .compressed_size = compressed_size,
+        .uncompressed_size = uncompressed_size,
+    };
+    const SAVEGAME_BSON_EXTENDED_HEADER extra_header = {
+        .flags = Game_GetBonusFlag(),
+        .counter = Savegame_GetCounter(),
+        .level_num = level->num,
+        .title_size = strlen(level->title),
+    };
+
+    File_WriteData(fp, &header, sizeof(header));
+    File_WriteData(fp, compressed, compressed_size);
+    File_WriteData(fp, &extra_header, sizeof(extra_header));
+    File_WriteData(fp, level->title, strlen(level->title));
+
+    Memory_FreePointer(&uncompressed);
+    Memory_FreePointer(&compressed);
+}
+
+static const char *M_GetSaveFilePattern(void)
+{
+    return g_GameFlow.savegame_fmt_bson;
+}
+
+static void M_SaveToFile(MYFILE *const fp, SAVEGAME_INFO *const info)
+{
+    const GF_LEVEL *const current_level = Game_GetCurrentLevel();
+    SAVEGAME_BSON_WRITE_CONTEXT *const ctx = Savegame_BSON_StartWrite();
+    JSON_VALUE *const root = Savegame_BSON_GetWriteRoot(ctx);
+
+    Savegame_BSON_DumpResumeInfoList(ctx);
+    Savegame_BSON_DumpInventory(ctx);
+    Savegame_BSON_DumpFlipmaps(ctx);
+    Savegame_BSON_DumpCameras(ctx);
+    Savegame_BSON_DumpItems(ctx);
+    Savegame_BSON_DumpEffects(ctx);
+    Savegame_BSON_DumpLara(ctx);
+    Savegame_BSON_DumpMusic(ctx);
+    Savegame_BSON_DumpFlares(ctx);
+    Savegame_BSON_DumpMisc(ctx);
+
+    M_SaveRaw(fp, root, current_level->num);
+    Savegame_BSON_FinishWrite(ctx);
+}
+
+static bool M_UpdateDeathCounters(
+    MYFILE *const fp, int32_t level_num, const int32_t death_count)
+{
+    bool result = false;
+    JSON_VALUE *const root = Savegame_BSON_ReadRaw(fp, nullptr);
+    JSON_OBJECT *const root_obj = JSON_ValueAsObject(root);
+    if (root_obj == nullptr) {
+        LOG_ERROR("Cannot find the root object");
+        goto cleanup;
+    }
+
+    JSON_OBJECT *const misc_obj = JSON_ObjectGetObject(root_obj, "misc");
+    if (misc_obj == nullptr) {
+        LOG_ERROR("Cannot find the misc object");
+        goto cleanup;
+    }
+    JSON_ObjectEvictKey(misc_obj, "death_count");
+    JSON_ObjectAppendInt(misc_obj, "death_count", death_count);
+
+    File_Seek(fp, 0, FILE_SEEK_SET);
+    M_SaveRaw(fp, root, level_num);
+    result = true;
+
+cleanup:
+    JSON_ValueFree(root);
+    return result;
+}
 
 static JSON_VALUE *M_ParseFromBuffer(
     const char *const buffer, int32_t *const version_out)
@@ -43,6 +159,64 @@ static JSON_VALUE *M_ParseFromBuffer(
     JSON_VALUE *const root = BSON_Parse(uncompressed, uncompressed_size);
     Memory_FreePointer(&uncompressed);
     return root;
+}
+
+static bool M_LoadFromFile(MYFILE *const fp)
+{
+    bool result = false;
+
+    int32_t version = -1;
+    JSON_VALUE *const root = Savegame_BSON_ReadRaw(fp, &version);
+    SAVEGAME_BSON_READ_CONTEXT *const ctx = Savegame_BSON_StartRead(root);
+
+    if (!Savegame_BSON_LoadMisc(ctx)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadResumeInfoList(ctx, version)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadInventory(ctx)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadFlipmaps(ctx)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadCameras(ctx)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadItems(ctx, version)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadEffects(ctx)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadFlares(ctx)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadMusic(ctx, version)) {
+        goto cleanup;
+    }
+    if (!Savegame_BSON_LoadLara(ctx, version)) {
+        goto cleanup;
+    }
+
+    result = true;
+
+cleanup:
+    Savegame_BSON_FinishRead(ctx, result);
+    JSON_ValueFree(root);
+    return result;
+}
+
+static bool M_LoadOnlyResumeInfo(MYFILE *const fp)
+{
+    int32_t version;
+    JSON_VALUE *const root = Savegame_BSON_ReadRaw(fp, &version);
+    SAVEGAME_BSON_READ_CONTEXT *const ctx = Savegame_BSON_StartRead(root);
+    const bool result = Savegame_BSON_LoadResumeInfoList(ctx, version);
+    Savegame_BSON_FinishRead(ctx, result);
+    JSON_ValueFree(root);
+    return result;
 }
 
 bool Savegame_BSON_FillInfo(MYFILE *const fp, SAVEGAME_INFO *const info)
@@ -97,60 +271,4 @@ JSON_VALUE *Savegame_BSON_ReadRaw(MYFILE *const fp, int32_t *const version_out)
     return result;
 }
 
-bool Savegame_BSON_LoadFromFile(MYFILE *const fp)
-{
-    bool result = false;
-
-    int32_t version = -1;
-    JSON_VALUE *const root = Savegame_BSON_ReadRaw(fp, &version);
-    SAVEGAME_BSON_READ_CONTEXT *const ctx = Savegame_BSON_StartRead(root);
-
-    if (!Savegame_BSON_LoadMisc(ctx)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadResumeInfoList(ctx, version)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadInventory(ctx)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadFlipmaps(ctx)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadCameras(ctx)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadItems(ctx, version)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadEffects(ctx)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadFlares(ctx)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadMusic(ctx, version)) {
-        goto cleanup;
-    }
-    if (!Savegame_BSON_LoadLara(ctx, version)) {
-        goto cleanup;
-    }
-
-    result = true;
-
-cleanup:
-    Savegame_BSON_FinishRead(ctx, result);
-    JSON_ValueFree(root);
-    return result;
-}
-
-bool Savegame_BSON_LoadOnlyResumeInfo(MYFILE *const fp)
-{
-    int32_t version;
-    JSON_VALUE *const root = Savegame_BSON_ReadRaw(fp, &version);
-    SAVEGAME_BSON_READ_CONTEXT *const ctx = Savegame_BSON_StartRead(root);
-    const bool result = Savegame_BSON_LoadResumeInfoList(ctx, version);
-    Savegame_BSON_FinishRead(ctx, result);
-    JSON_ValueFree(root);
-    return result;
-}
+REGISTER_SAVEGAME_STRATEGY(m_Strategy)
